Added calls to SmsService and InstanceUploader to not defer/defer submission states.
SmsService now flags a submission as not deferred when it's being made so that the SMS status will be given priority in the list. Since InstanceUploader is primarily being used by Google Sheets and Aggregate then it was chosen as the centralized point to flag a SMS submission as deferred if it exists.